+++ bepusdt/__init__.py
@@ -1,10 +1,18 @@
 """BEpusdt Python SDK - USDT/TRX/USDC 支付网关客户端"""
 
 from .client import BEpusdtClient
-from .exceptions import BEpusdtError, SignatureError, APIError
+from .exceptions import (
+    BEpusdtError, SignatureError, APIError,
+    NetworkError, TimeoutError, ServerError, ClientError, ValidationError
+)
 from .models import Order, OrderStatus, TradeType
 
-__version__ = "0.2.1"
+__version__ = "0.2.2"
 __author__ = "luoyanglang"
 __url__ = "https://github.com/luoyanglang/bepusdt-python-sdk"
-__all__ = ["BEpusdtClient", "BEpusdtError", "SignatureError", "APIError", "Order", "OrderStatus", "TradeType"]
+__all__ = [
+    "BEpusdtClient",
+    "BEpusdtError", "SignatureError", "APIError",
+    "NetworkError", "TimeoutError", "ServerError", "ClientError", "ValidationError",
+    "Order", "OrderStatus", "TradeType"
+]